added utils file to contain all shared utils
